seperated the graphs into components so that i can render certain graphs for certian components
