+++ frontend/src/containers/Home.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import AdminContainer from './AdminContainer'
-import HomeGraph from '../components/HomeGraph'
+import HomeGraph from './HomeGraph'
 import ManagerContainer from './ManagerContainer';
 import DeveloperContainer from './DeveloperContainer';
 import SubmitterContainer from './SubmitterContainer';
